submodule--helper: convert the bulk of cmd_add() to C
Introduce the 'add' subcommand to submodule--helper that does all the
work 'submodule add' past the parsing of flags.

While at it, remove the now-unused shell interface to
'resolve_relative_url()' and repurpose it as an internal helper
function.

Since the 'die()' of C prepends the word 'fatal', we modify t7400.6 to
reflect the same.

Signed-off-by: Atharva Raykar <raykar.ath@gmail.com>
Mentored-by: Christian Couder <christian.couder@gmail.com>
Helped-by: Kaartic Sivaraam <kaartic.sivaraam@gmail.com>
Mentored-by: Shourya Shukla <periperidip@gmail.com>