submodule--helper: introduce internal helper
Add a helper function to resolve a relative url, that is meant to be
called from within C, thus not needing any flag parsing logic.

This function will be useful in the conversion that is to follow, and
will replace `resolve_relative_url`, when that is not needed anymore.

This function behaves the same as `resolve_relative_url` when a single
argument is supplied to it (and on further inspection, it is never
called anywhere with more than one argument). There is some repetition
of code happening here, but it is addressed in a later commit, which
will remove the shell interface to `resolve_relative_url` along with the
reduncant code.

Signed-off-by: Atharva Raykar <raykar.ath@gmail.com>
Mentored-by: Christian Couder <christian.couder@gmail.com>
Mentored-by: Shourya Shukla <shouryashukla.oo@gmail.com>