submodule--helper: get remote names from any repository
`get_default_remote()` retrieves the name of a remote by resolving the
refs from of the current repository's ref store.

Thus in order to use it for retrieving the remote name of a submodule,
we have to start a new process which runs from the submodule directory.

Let's instead split off the function to handle two cases. To use it for
retrieving a submodule's remote, we use `get_default_remote_submodule()`
which additionally takes the path to the submodule.

`get_default_remote()` retrieves the remote for the current repository,
ie, it's behaviour remains unchanged.

Both of these functions in turn call `repo_get_default_remote()` which
takes any repository object and retrieves the remote accordingly.

Signed-off-by: Atharva Raykar <raykar.ath@gmail.com>
Mentored-by: Christian Couder <christian.couder@gmail.com>
Mentored-by: Shourya Shukla <periperidip@gmail.com>