submodule--helper: make struct update_data clone-aware
We add more fields to the `struct update_data` that are required by
`struct submodule_update_clone` to perform a clone.

This will be of use to us in the next patch, where we use this struct as
the main object that holds all the data needed for the `submodule
update` command.

Signed-off-by: Atharva Raykar <raykar.ath@gmail.com>
Mentored-by: Christian Couder <christian.couder@gmail.com>
Mentored-by: Shourya Shukla <periperidip@gmail.com>