submodule: create skeleton for builtin
Now that most of the core logic of 'git-submodule.sh' is handled in C,
let's start converting it to a true builtin. We mirror similar such
efforts of migrating shell scripts to builtin in the past, by:

- Renaming git-submodule to git-legacy-submodule.
- Creating a submodule.useBuiltin switch. When this is set to false, our
  builtin just exec()'s the legacy script.

On this skeleton, we will start copying the functions of
'submodule--helper.c', to this file, while adapting the flag parsing to
match what is being done in the legacy shell script.

Once we the transition to the builtin is complete we will set the
default value of 'submodule.useBuiltin=true'. This lets us provide an
escape hatch for the users if they discover a bug in the builtin.

Once we are sure of the new builtin's reliability, we will remove this
switch entirely, along with the old submodule--helper subcommand and the
legacy script.

Signed-off-by: Atharva Raykar <raykar.ath@gmail.com>
Mentored-by: Christian Couder <christian.couder@gmail.com>
Mentored-by: Shourya Shukla <periperidip@gmail.com>