tests: add a special setup where submodule.useBuiltin is off
Add a GIT_TEST_SUBMODULE_USE_BUILTIN=false test mode which is equivalent
to running with submodule.useBuiltin=false. This is needed to spot that
we're not introducing any regressions in the legacy submodule version
while we're carrying both it and the new built-in version.

This imitates the equivalent treatment for the built-in rebase in
62c23938fae5 (tests: add a special setup where rebase.useBuiltin is off,
2018-11-14).

Signed-off-by: Atharva Raykar <raykar.ath@gmail.com>
Mentored-by: Christian Couder <christian.couder@gmail.com>
Mentored-by: Shourya Shukla <periperidip@gmail.com>