submodule: convert status to C builtin
Let's make 'submodule status' to a C builtin. We do this by copying
verbatim all the necessary functions that have already implemented
'submodule status' from 'submodule--helper.c'.

Even the invocation of 'git submodule' which defaults to the 'status'
subcommand has been handled in this patch.

The only truly new code this patch introduces is in 'cmd_submodule()'.
We also create a helper function that maintains all the subcommands that
have been converted to C builtins. This lets us detect when the user
requests a command that has not been made a builtin yet, and makes Git
fall back to the legacy script.

This mechanism is particularly useful for satisfying the test suite
until a full conversion is done.

Signed-off-by: Atharva Raykar <raykar.ath@gmail.com>
Mentored-by: Christian Couder <christian.couder@gmail.com>
Mentored-by: Shourya Shukla <periperidip@gmail.com>